2.6.0 improve namespacing and classloading
